+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_document_analysis_client.py
@@ -84,8 +84,8 @@ class DocumentAnalysisClient(FormRecognizerClientBase):
             `pages="1-3, 5-6"`. Separate each page number or range with a comma.
         :keyword str locale: Locale hint of the input document.
             See supported locales here: https://aka.ms/azsdk/formrecognizer/supportedlocales.
-        :keyword list[str] features: Document analysis features to enable. See supported
-            analysis features on :class:`~azure.ai.formrecognizer.AnalysisFeature`.
+        :keyword features: Document analysis features to enable.
+        :paramtype: list[str or ~azure.ai.formrecognizer.AnalysisFeature]
         :return: An instance of an LROPoller. Call `result()` on the poller
             object to return a :class:`~azure.ai.formrecognizer.AnalyzeResult`.
         :rtype: ~azure.core.polling.LROPoller[~azure.ai.formrecognizer.AnalyzeResult]
@@ -158,8 +158,8 @@ class DocumentAnalysisClient(FormRecognizerClientBase):
             `pages="1-3, 5-6"`. Separate each page number or range with a comma.
         :keyword str locale: Locale hint of the input document.
             See supported locales here: https://aka.ms/azsdk/formrecognizer/supportedlocales.
-        :keyword list[str] features: Document analysis features to enable. See supported
-            analysis features on :class:`~azure.ai.formrecognizer.AnalysisFeature`.
+        :keyword features: Document analysis features to enable.
+        :paramtype: list[str or ~azure.ai.formrecognizer.AnalysisFeature]
         :return: An instance of an LROPoller. Call `result()` on the poller
             object to return a :class:`~azure.ai.formrecognizer.AnalyzeResult`.
         :rtype: ~azure.core.polling.LROPoller[~azure.ai.formrecognizer.AnalyzeResult]