+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_document_analysis_client_async.py
@@ -87,8 +87,8 @@ class DocumentAnalysisClient(FormRecognizerClientBaseAsync):
             `pages="1-3, 5-6"`. Separate each page number or range with a comma.
         :keyword str locale: Locale hint of the input document.
             See supported locales here: https://aka.ms/azsdk/formrecognizer/supportedlocales.
-        :keyword list[str] features: Document analysis features to enable. See supported
-            analysis features on :class:`~azure.ai.formrecognizer.AnalysisFeature`.
+        :keyword features: Document analysis features to enable.
+        :paramtype: list[str or ~azure.ai.formrecognizer.AnalysisFeature]
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a :class:`~azure.ai.formrecognizer.AnalyzeResult`.
         :rtype: ~azure.core.polling.AsyncLROPoller[~azure.ai.formrecognizer.AnalyzeResult]
@@ -161,8 +161,8 @@ class DocumentAnalysisClient(FormRecognizerClientBaseAsync):
             `pages="1-3, 5-6"`. Separate each page number or range with a comma.
         :keyword str locale: Locale hint of the input document.
             See supported locales here: https://aka.ms/azsdk/formrecognizer/supportedlocales.
-        :keyword list[str] features: Document analysis features to enable. See supported
-            analysis features on :class:`~azure.ai.formrecognizer.AnalysisFeature`.
+        :keyword features: Document analysis features to enable.
+        :paramtype: list[str or ~azure.ai.formrecognizer.AnalysisFeature]
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a :class:`~azure.ai.formrecognizer.AnalyzeResult`.
         :rtype: ~azure.core.polling.AsyncLROPoller[~azure.ai.formrecognizer.AnalyzeResult]