+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -3,8 +3,8 @@
 ## 3.3.0b1 (Unreleased)
 
 ### Features Added
-- Added new `features` kwarg on `begin_analyze_document()` and `begin_analyze_document_from_url()`.
-- Added new `AnalysisFeature` enum with optional document analysis feature to enable.
+- Added `features` keyword argument on `begin_analyze_document()` and `begin_analyze_document_from_url()`.
+- Added `AnalysisFeature` enum with optional document analysis feature to enable.
 - Added the following optional properties on `DocumentStyle` class: `similar_font_family`, `font_style`, `font_weight`, `color`, `background_color`.
 
 ### Breaking Changes